Up the log like float size

+++ src/pigglet/constants.py
@@ -7,6 +7,6 @@ HET_NUM = 1
 HOM_REF_NUM = 0
 ROOT_LABEL = -1
 TMP_LABEL = -2
-LOG_LIKE_DTYPE = np.float32
+LOG_LIKE_DTYPE = np.float64
 GL_DTYPE = np.float32
 REAL_SPACE_LIKE_DTYPE = np.float32
